# src/main/java/com/ibm/api/BqPaymentExecutionWorkstepApi.java
/**
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (7.2.0).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package com.ibm.api;

import com.ibm.model.HTTPError;
import com.ibm.model.RetrievePaymentExecutionWorkstepResponse;
import io.swagger.annotations.*;
import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.enums.ParameterIn;
import org.springframework.http.HttpStatus;
import org.springframework.http.MediaType;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.PathVariable;
import org.springframework.web.bind.annotation.RequestBody;
import org.springframework.web.bind.annotation.RequestHeader;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RequestMethod;
import org.springframework.web.bind.annotation.RequestParam;
import org.springframework.web.bind.annotation.RequestPart;
import org.springframework.web.context.request.NativeWebRequest;
import org.springframework.web.multipart.MultipartFile;

import javax.validation.Valid;
import javax.validation.constraints.*;
import java.util.List;
import java.util.Map;
import java.util.Optional;
@Validated
@Api(value = "BqPaymentExecutionWorkstep", description = "the BqPaymentExecutionWorkstep API")
public interface BqPaymentExecutionWorkstepApi {
    

  @ApiOperation(value = "ReBQ Retrieve details about the gateway transaction", nickname = "retrievePaymentExecutionWorkstep", notes = "ReBQ Retrieve details about the gateway transaction", response = RetrievePaymentExecutionWorkstepResponse.class , tags = { "BQ - PaymentExecutionWorkstep" }  )
    @ApiResponses({ @ApiResponse(code = 200, message = "RetrievePaymentExecutionWorkstepResponse", response = RetrievePaymentExecutionWorkstepResponse.class),@ApiResponse(code = 400, message = "BadRequest", response = HTTPError.class),@ApiResponse(code = 401, message = "Unauthorized", response = HTTPError.class),@ApiResponse(code = 403, message = "Forbidden", response = HTTPError.class),@ApiResponse(code = 404, message = "NotFound", response = HTTPError.class),@ApiResponse(code = 429, message = "TooManyRequests", response = HTTPError.class),@ApiResponse(code = 500, message = "InternalServerError", response = HTTPError.class) })
     @RequestMapping(method = RequestMethod.GET,
            value = "/PaymentExecution/{paymentexecutionid}/PaymentExecutionWorkstep/{paymentexecutionworkstepid}/Retrieve",
            produces = { "application/json" })
    ResponseEntity<RetrievePaymentExecutionWorkstepResponse> retrievePaymentExecutionWorkstep(@Parameter(name = "paymentexecutionid", description = "", required = true, in = ParameterIn.PATH) @PathVariable("paymentexecutionid") String paymentexecutionid,@Parameter(name = "paymentexecutionworkstepid", description = "", required = true, in = ParameterIn.PATH) @PathVariable("paymentexecutionworkstepid") String paymentexecutionworkstepid);
      
}