# src/main/resources/openapi.yaml
openapi: 3.0.1
info:
  contact:
    name: BIAN
    url: https://bian.org
  description: "This service domains processes the movement of funds between accounts\
    \ within the bank or to/from an account held with another bank, selecting the\
    \ appropriate payment mechanism"
  title: Payment Execution
  version: 11.0.0
servers:
- url: https://virtserver.swaggerhub.com/BIAN-3/BIAN/PaymentExecution/11.0.0
paths:
...
  /PaymentExecution/{paymentexecutionid}/Retrieve:
    get:
      description: ReCR Retrieve details about a payment execution instruction
      operationId: Retrieve
      parameters:
      - explode: false
        in: path
        name: paymentexecutionid
        required: true
        schema:
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RetrievePaymentExecutionProcedureResponse'
          description: RetrievePaymentExecutionProcedureResponse
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
          description: BadRequest
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
          description: Unauthorized
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
          description: Forbidden
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
          description: NotFound
        "429":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
          description: TooManyRequests
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
          description: InternalServerError
      summary: ReCR Retrieve details about a payment execution instruction
      tags:
      - CR - PaymentExecutionProcedure
      x-accepts: application/json
      x-tags:
      - tag: CR - PaymentExecutionProcedure
  /PaymentExecution/{paymentexecutionid}/PaymentMechanism/{paymentmechanismid}/Retrieve:
    get:
      description: ReBQ Retrieve details about the payment mechanism selection
      operationId: RetrievePaymentMechanism
      parameters:
      - explode: false
        in: path
        name: paymentexecutionid
        required: true
        schema:
          type: string
        style: simple
      - explode: false
        in: path
        name: paymentmechanismid
        required: true
        schema:
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RetrievePaymentMechanismResponse'
          description: RetrievePaymentMechanismResponse
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
          description: BadRequest
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
          description: Unauthorized
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
          description: Forbidden
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
          description: NotFound
        "429":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
          description: TooManyRequests
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
          description: InternalServerError
      summary: ReBQ Retrieve details about the payment mechanism selection
      tags:
      - BQ - PaymentMechanism
      x-accepts: application/json
      x-tags:
      - tag: BQ - PaymentMechanism
  /PaymentExecution/{paymentexecutionid}/PaymentExecutionWorkstep/{paymentexecutionworkstepid}/Retrieve:
    get:
      description: ReBQ Retrieve details about the gateway transaction
      operationId: RetrievePaymentExecutionWorkstep
      parameters:
      - explode: false
        in: path
        name: paymentexecutionid
        required: true
        schema:
          type: string
        style: simple
      - explode: false
        in: path
        name: paymentexecutionworkstepid
        required: true
        schema:
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RetrievePaymentExecutionWorkstepResponse'
          description: RetrievePaymentExecutionWorkstepResponse
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
          description: BadRequest
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
          description: Unauthorized
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
          description: Forbidden
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
          description: NotFound
        "429":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
          description: TooManyRequests
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
          description: InternalServerError
      summary: ReBQ Retrieve details about the gateway transaction
      tags:
      - BQ - PaymentExecutionWorkstep
      x-accepts: application/json
      x-tags:
      - tag: BQ - PaymentExecutionWorkstep
components:
...
  schemas:
    HTTPError:
      description: Status Attributes
      properties:
        status_code:
          format: string
          type: string
        status:
          format: string
          type: string
        message:
          format: string
          type: string
      type: object
...
    RetrievePaymentExecutionProcedureResponse:
      description: "Output: ReCR Retrieve details about a payment execution instruction"
      example:
        PaymentExecutionProcedure:
          PayeeProductInstanceReference: PayeeProductInstanceReference
          WashAccountReference: WashAccountReference
          Amount: Amount
          CorrespondentServiceAccessSchedule: CorrespondentServiceAccessSchedule
          ACHAccessSchedule: ACHAccessSchedule
          PaymentMechanismType: PaymentMechanismType
          PayerBankReference: PayerBankReference
          PaymentOrderReference: PaymentOrderReference
          PayeeReference: PayeeReference
          IntermediaryAccountReference: IntermediaryAccountReference
          PayerReference: PayerReference
          ACHAccessServiceSessionReference: ACHAccessServiceSessionReference
          PaymentTransactionInitiatorReference: PaymentTransactionInitiatorReference
          PayeeBankReference: PayeeBankReference
          Currency: Currency
          PayerProductInstanceReference: PayerProductInstanceReference
          DateType: DateType
          CorrespondentServiceSessionReference: CorrespondentServiceSessionReference
      properties:
        PaymentExecutionProcedure:
          $ref: '#/components/schemas/RetrievePaymentExecutionProcedureResponse_PaymentExecutionProcedure'
      type: object
    RetrievePaymentMechanismResponse:
      description: "Output: ReBQ Retrieve details about the payment mechanism selection"
      example:
        PaymentMechanism:
          PaymentExecutionTaskResult: PaymentExecutionTaskResult
        PaymentExecutionProcedure:
          PayeeProductInstanceReference: PayeeProductInstanceReference
          WashAccountReference: WashAccountReference
          Amount: Amount
          CorrespondentServiceAccessSchedule: CorrespondentServiceAccessSchedule
          ACHAccessSchedule: ACHAccessSchedule
          PaymentMechanismType: PaymentMechanismType
          PayerBankReference: PayerBankReference
          PaymentOrderReference: PaymentOrderReference
          PayeeReference: PayeeReference
          IntermediaryAccountReference: IntermediaryAccountReference
          PayerReference: PayerReference
          ACHAccessServiceSessionReference: ACHAccessServiceSessionReference
          PaymentTransactionInitiatorReference: PaymentTransactionInitiatorReference
          PayeeBankReference: PayeeBankReference
          Currency: Currency
          PayerProductInstanceReference: PayerProductInstanceReference
          DateType: DateType
          CorrespondentServiceSessionReference: CorrespondentServiceSessionReference
      properties:
        PaymentExecutionProcedure:
          $ref: '#/components/schemas/RetrievePaymentExecutionProcedureResponse_PaymentExecutionProcedure'
        PaymentMechanism:
          $ref: '#/components/schemas/RetrievePaymentMechanismResponse_PaymentMechanism'
      type: object
    RetrievePaymentExecutionWorkstepResponse:
      description: "Output: ReBQ Retrieve details about the gateway transaction"
      example:
        PaymentExecutionProcedure:
          PayeeProductInstanceReference: PayeeProductInstanceReference
          WashAccountReference: WashAccountReference
          Amount: Amount
          CorrespondentServiceAccessSchedule: CorrespondentServiceAccessSchedule
          ACHAccessSchedule: ACHAccessSchedule
          PaymentMechanismType: PaymentMechanismType
          PayerBankReference: PayerBankReference
          PaymentOrderReference: PaymentOrderReference
          PayeeReference: PayeeReference
          IntermediaryAccountReference: IntermediaryAccountReference
          PayerReference: PayerReference
          ACHAccessServiceSessionReference: ACHAccessServiceSessionReference
          PaymentTransactionInitiatorReference: PaymentTransactionInitiatorReference
          PayeeBankReference: PayeeBankReference
          Currency: Currency
          PayerProductInstanceReference: PayerProductInstanceReference
          DateType: DateType
          CorrespondentServiceSessionReference: CorrespondentServiceSessionReference
        PaymentExecutionWorkstep:
          PaymentExecutionTaskResult: PaymentExecutionTaskResult
      properties:
        PaymentExecutionProcedure:
          $ref: '#/components/schemas/RetrievePaymentExecutionProcedureResponse_PaymentExecutionProcedure'
        PaymentExecutionWorkstep:
          $ref: '#/components/schemas/RetrievePaymentMechanismResponse_PaymentMechanism'
      type: object
...
    RetrievePaymentExecutionProcedureResponse_PaymentExecutionProcedure:
      example:
        PayeeProductInstanceReference: PayeeProductInstanceReference
        WashAccountReference: WashAccountReference
        Amount: Amount
        CorrespondentServiceAccessSchedule: CorrespondentServiceAccessSchedule
        ACHAccessSchedule: ACHAccessSchedule
        PaymentMechanismType: PaymentMechanismType
        PayerBankReference: PayerBankReference
        PaymentOrderReference: PaymentOrderReference
        PayeeReference: PayeeReference
        IntermediaryAccountReference: IntermediaryAccountReference
        PayerReference: PayerReference
        ACHAccessServiceSessionReference: ACHAccessServiceSessionReference
        PaymentTransactionInitiatorReference: PaymentTransactionInitiatorReference
        PayeeBankReference: PayeeBankReference
        Currency: Currency
        PayerProductInstanceReference: PayerProductInstanceReference
        DateType: DateType
        CorrespondentServiceSessionReference: CorrespondentServiceSessionReference
      properties:
        ACHAccessServiceSessionReference:
          description: |2+

          format: object
          type: string
        ACHAccessSchedule:
          description: |2+

          format: string
          type: string
        CorrespondentServiceSessionReference:
          description: |2+

          format: object
          type: string
        CorrespondentServiceAccessSchedule:
          description: |2+

          format: string
          type: string
        PaymentTransactionInitiatorReference:
          description: |2+

          format: object
          type: string
        PaymentOrderReference:
          description: |2+

          format: object
          type: string
        PayerReference:
          description: |2+

          format: object
          type: string
        PayerBankReference:
          description: |2+

          format: object
          type: string
        PayerProductInstanceReference:
          description: |2+

          format: object
          type: string
        PayeeReference:
          description: |2+

          format: object
          type: string
        PayeeBankReference:
          description: |2+

          format: object
          type: string
        PayeeProductInstanceReference:
          description: |2+

          format: object
          type: string
        Amount:
          description: |2+

          format: string
          type: string
        Currency:
          description: |2+

          format: string
          type: string
        DateType:
          description: |2+

          format: string
          type: string
        WashAccountReference:
          description: |2+

          format: object
          type: string
        IntermediaryAccountReference:
          description: |2+

          format: object
          type: string
        PaymentMechanismType:
          description: |2+

          format: string
          type: string
      type: object
    RetrievePaymentMechanismResponse_PaymentMechanism:
      example:
        PaymentExecutionTaskResult: PaymentExecutionTaskResult
      properties:
        PaymentExecutionTaskResult:
          description: |2+

          format: string
          type: string
      type: object